Fix collector close issue and add perf test for multiple exporters to collector
